Adicionando a parte de vendas e clientes, adicionando o .gitignore.

+++ Models/Estoque.py
@@ -13,8 +13,16 @@ class Estoque:
         self.conn.commit()
 
     def remover_produto(self, codigo):  # remove um produto pelo seu código
-        self.cursor.execute("DELETE FROM estoque WHERE codigo = ?", (codigo,))
-        self.conn.commit()  # Confirma a transação no banco de dados
+        self.cursor.execute("SELECT * FROM estoque WHERE codigo = ?", (codigo,))
+        resultado = self.cursor.fetchall()
+        if resultado == []:
+            print("Produto nao encontrado")
+            sleep(1)
+        else:
+            self.cursor.execute("DELETE FROM estoque WHERE codigo = ?", (codigo,))
+            self.conn.commit()  # Confirma a transação no banco de dados
+            print("Produto removido com sucesso!")
+            sleep(1)
 
     def consultar_produto(self, codigo): #consulta um produto pelo seu código
         self.cursor.execute("SELECT * FROM estoque WHERE codigo = ?", (codigo,))
@@ -37,4 +45,28 @@ class Estoque:
             input("Pressione ENTER para voltar: ")
         else:
             print("O estoque está vazio.")
+            input("Pressione ENTER para voltar: ")
+    
+    def verificar_estoque(self, codigo):
+        self.cursor.execute("SELECT codigo FROM estoque WHERE codigo = ?", (codigo,))
+        produto = self.cursor.fetchone()
+        if produto:
+            return True
+        else:
+            print("Produto não encontrado.")
+            input("Pressione ENTER para voltar: ")
+            return False
+    def atualizar_estoque(self, codigo, quantidade):
+        self.cursor.execute("SELECT quantidade FROM estoque WHERE codigo = ?", (codigo,))
+        quantidade_no_estoque = self.cursor.fetchone()[0]
+        if quantidade_no_estoque < int(quantidade):
+            print("Quantidade insuficiente no estoque.")
+            sleep(1)
+            return False
+        else:
+            self.cursor.execute("UPDATE estoque SET quantidade = ? WHERE codigo = ?", (quantidade_no_estoque - int(quantidade), codigo))
+            self.conn.commit()
+            print("Estoque atualizado!")
+            sleep(1)
+            return True
